[BUG] training sorting not working correctly
After changing the url structure from /courses to /training, the js
ajax calls were not updated properly and the sorting functions
weren't working.

This commit also changes to use the index of the class instead of the
order for display, as the order can skip numbers if there are draft
lessons.