[BUG] Fix training lessons showing unpublished
This commit fixes a few issues:
 First, if the lesson_order isn't sequential for some reason, the next
lesson function didn't work properly.  That function is a bit more
robust now, and will just return the next (published) lesson with a
heigher lesson_order.

Second, in not all areas was the status of the lesson checked, so
draft lessons were being shown.