[BUG] First pass at fixing the Font issue...
The site uses Open Sans by default, but it was never being loaded. So,
in reality, each browser was falling back to a system font, which
displays differently everywhere, causing layout bugs.  I now load
Open Sans, though we only use it in one bug-prone header spot.

In the future we'll be moving fully to Open Sans, but need to check
each location for layout breaks when doing so.